requirements.txt
tkinter
pillow
opencv-python
